requirements.txt
absl_py==2.1.0
deep_sort_realtime==1.3.2
easydict==1.13
ipdb==0.13.13
matplotlib==3.9.0
motmetrics==1.4.0
numpy==1.23.5
opencv_python==4.10.0.82
pandas==2.2.2
Pillow==10.4.0
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.14.0
streamlit==1.36.0
torch==2.0.1+cu117
torchvision==0.15.2+cu117
ultralytics==8.1.34
